# Remove extraneous clones in TypeNamed and ViolationChild